+++ sdk/ui-vue/src/components/PlAgDataTable/sources/table-source.ts
@@ -3,7 +3,6 @@ import type { AxisId, JoinEntry, PColumnIdAndSpec, PFrameHandle, PObjectId } fro
 import {
   type PColumnSpec,
   type PFrameDriver,
-  type PTableColumnId,
   type PTableColumnSpec,
   type PTableHandle,
   type PTableVector,
@@ -23,17 +22,14 @@ import { getHeterogeneousColumns, updatePFrameGridOptionsHeterogeneousAxes } fro
  * Generate unique colId based on the column spec.
  */
 function makeColId(spec: PTableColumnSpec) {
-  return canonicalize({
-    type: spec.type,
-    id: spec.id,
-  })!;
+  return canonicalize(spec)!;
 }
 
 /**
  * Extract `PTableColumnId` from colId string
  */
 export function parseColId(str: string) {
-  return JSON.parse(str) as PTableColumnId;
+  return JSON.parse(str) as PTableColumnSpec;
 }
 
 // do not use `any` please
@@ -231,6 +227,12 @@ export async function makeSheets(
   });
 }
 
+export type PlAgDataTableRow = {
+  id: string;
+  key: unknown[];
+  [field: `${number}`]: undefined | null | number | string;
+};
+
 /**
  * Convert columnar data from the driver to rows, used by ag-grid
  * @param specs column specs
@@ -238,13 +240,16 @@ export async function makeSheets(
  * @param nRows number of rows
  * @returns
  */
-function columns2rows(fields: number[], columns: PTableVector[], index: number): unknown[] {
-  const nCols = columns.length;
-  const rowData = [];
+function columns2rows(fields: number[], columns: PTableVector[], axes: number[], index: number): PlAgDataTableRow[] {
+  const nCols = fields.length;
+  const rowData: PlAgDataTableRow[] = [];
   for (let iRow = 0; iRow < columns[0].data.length; ++iRow) {
-    const row: Record<string, unknown> = {};
+    const key: unknown[] = [];
+    const id = (index++).toString();
+    for (const axisIdx of axes) key.push(columns[axisIdx].data[iRow]);
+    const row: PlAgDataTableRow = { key, id };
     for (let iCol = 0; iCol < nCols; ++iCol) {
-      const field = fields[iCol].toString();
+      const field = fields[iCol].toString() as `${number}`;
       const value = columns[iCol].data[iRow];
       const valueType = columns[iCol].type;
       if (isValueAbsent(columns[iCol].absent, iRow)) {
@@ -255,7 +260,7 @@ function columns2rows(fields: number[], columns: PTableVector[], index: number):
         row[field] = toDisplayValue(value, valueType);
       }
     }
-    row['id'] = (index++).toString();
+
     rowData.push(row);
   }
   return rowData;
@@ -278,6 +283,9 @@ export async function updatePFrameGridOptions(
   rowData?: unknown[];
 }> {
   const specs = await pfDriver.getSpec(pt);
+
+  let numberOfAxes = specs.findIndex((s) => s.type === 'column');
+  if (numberOfAxes === -1) numberOfAxes = specs.length;
 
   // column indices in the specs array that we are going to process
   const indices = [...specs.keys()]
@@ -333,6 +341,36 @@ export async function updatePFrameGridOptions(
     return updatePFrameGridOptionsHeterogeneousAxes(hColumns, ptShape, columnDefs, await pfDriver.getData(pt, indices), fields, indices);
   }
 
+  // mixing in axis indices
+
+  const allIndices = [...indices];
+
+  // axisIdx (0..<axesCount) -> idx in allIndices array
+  const axisToFieldIdx = new Map<number, number>();
+  for (let i = 0; i < numberOfAxes; ++i) axisToFieldIdx.set(i, -1);
+
+  for (let i = 0; i < allIndices.length; ++i) {
+    const idx = allIndices[i];
+    if (axisToFieldIdx.has(idx)) axisToFieldIdx.set(idx, i);
+  }
+
+  // at this point we have axis indices that are not listed in indices set to -1 in axisToFieldIdx
+
+  // adding those indices at the end of allIndices array, to make sure we have all the axes in our response
+  for (const [key, value] of axisToFieldIdx)
+    if (value === -1) {
+      axisToFieldIdx.set(key, allIndices.length /* at this index value will be inserted in the next line */);
+      allIndices.push(key);
+    }
+
+  // indices of axes in allIndices array
+  const axes: number[] = [];
+  for (let i = 0; i < numberOfAxes; ++i) {
+    const fieldIdx = axisToFieldIdx.get(i);
+    if (fieldIdx === undefined || fieldIdx === -1) throw new Error('assertion exception');
+    axes.push(fieldIdx);
+  }
+
   let lastParams: IServerSideGetRowsParams | undefined = undefined;
   const serverSideDatasource = {
     getRows: async (params: IServerSideGetRowsParams) => {
@@ -347,8 +385,8 @@ export async function updatePFrameGridOptions(
         // this is to avoid double flickering when underlying table is changed
         if (lastParams && !lodash.isEqual(lastParams.request.sortModel, params.request.sortModel)) {
           lastParams = undefined;
-          params.fail();
           params.api.setGridOption('loading', true);
+          params.success({ rowData: [], rowCount: 0 });
           return;
         }
         lastParams = params;
@@ -358,11 +396,11 @@ export async function updatePFrameGridOptions(
         if (rowCount > 0 && params.request.startRow !== undefined && params.request.endRow !== undefined) {
           length = Math.min(rowCount, params.request.endRow) - params.request.startRow;
           if (length > 0) {
-            const data = await pfDriver.getData(pt, indices, {
+            const data = await pfDriver.getData(pt, allIndices, {
               offset: params.request.startRow,
               length,
             });
-            rowData = columns2rows(fields, data, params.request.startRow);
+            rowData = columns2rows(fields, data, axes, params.request.startRow);
           }
         }
 
@@ -370,8 +408,9 @@ export async function updatePFrameGridOptions(
         params.api.autoSizeAllColumns();
         params.api.setGridOption('loading', false);
       } catch (error: unknown) {
-        params.fail();
         params.api.setGridOption('loading', true);
+        params.fail();
+        console.trace(error);
       }
     },
   } satisfies IServerSideDatasource;