+++ etc/block-components/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies [83caf43]
+  - @platforma-sdk/ui-vue@1.10.4
+
+## 1.2.11
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/ui-vue@1.10.2
+  - @platforma-sdk/model@1.10.2
+  - @milaboratories/milaboratories.ui-examples.model@1.0.35
+
+## 1.2.10
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+  - @milaboratories/milaboratories.ui-examples.model@1.0.34
+  - @platforma-sdk/ui-vue@1.10.0
+
+## 1.2.9
+
+### Patch Changes
+
+- Updated dependencies [1f12f1a]
+  - @platforma-sdk/ui-vue@1.9.3
+
 ## 1.2.8
 
 ### Patch Changes