+++ etc/blocks/enter-numbers/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.86
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.86
+
+## 1.0.85
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/model@1.10.2
+  - @milaboratories/milaboratories.test-enter-numbers.ui@1.0.85
+  - @milaboratories/milaboratories.test-enter-numbers.model@1.0.21
+
+## 1.0.84
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+  - @milaboratories/milaboratories.test-enter-numbers.model@1.0.20
+  - @milaboratories/milaboratories.test-enter-numbers.ui@1.0.84
+
+## 1.0.83
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.83
+
 ## 1.0.82
 
 ### Patch Changes