+++ etc/blocks/model-test/block/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.test-block-model
 
+## 1.0.5
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/model@1.10.2
+  - @milaboratories/milaboratories.test-block-model.ui@1.0.4
+  - @milaboratories/milaboratories.test-block-model.model@1.0.3
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+  - @milaboratories/milaboratories.test-block-model.model@1.0.2
+  - @milaboratories/milaboratories.test-block-model.ui@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes