+++ etc/blocks/model-test/model/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/milaboratories.test-block-model.model
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/model@1.10.2
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+
 ## 1.0.1
 
 ### Patch Changes