+++ etc/blocks/model-test/ui/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @milaboratories/milaboratories.test-block-model.ui
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [83caf43]
+  - @platforma-sdk/ui-vue@1.10.4
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/ui-vue@1.10.2
+  - @milaboratories/milaboratories.test-block-model.model@1.0.3
+
+## 1.0.3
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.model@1.0.2
+- @platforma-sdk/ui-vue@1.10.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [1f12f1a]
+  - @platforma-sdk/ui-vue@1.9.3
+
 ## 1.0.1
 
 ### Patch Changes