+++ etc/blocks/ui-examples/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.92
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.ui@1.2.12
+
+## 1.0.91
+
+### Patch Changes
+
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/model@1.10.2
+  - @milaboratories/milaboratories.ui-examples.ui@1.2.11
+  - @milaboratories/milaboratories.ui-examples.model@1.0.35
+
+## 1.0.90
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+  - @milaboratories/milaboratories.ui-examples.model@1.0.34
+  - @milaboratories/milaboratories.ui-examples.ui@1.2.10
+
+## 1.0.89
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.ui@1.2.9
+
 ## 1.0.88
 
 ### Patch Changes