+++ sdk/ui-vue/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @platforma-sdk/ui-vue
 
+## 1.10.4
+
+### Patch Changes
+
+- 83caf43: PFrames version bump, enabled fixed filters
+
+## 1.10.2
+
+### Patch Changes
+
+- 8187ba2: PlAgDataTable filters implementation
+- Updated dependencies [8187ba2]
+  - @platforma-sdk/model@1.10.2
+  - @milaboratories/uikit@2.2.3
+
+## 1.10.0
+
+### Patch Changes
+
+- Updated dependencies [b089273]
+  - @platforma-sdk/model@1.10.0
+  - @milaboratories/uikit@2.2.3
+
+## 1.9.3
+
+### Patch Changes
+
+- 1f12f1a: Added onRowDoubleClicked emit for PlAgDataTable
+
 ## 1.9.0
 
 ### Patch Changes