+++ tools/tengo-builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-sdk/tengo-builder
 
+## 1.17.2
+
+### Patch Changes
+
+- 7034bdb: Produce correct library import names when facing deep index.lib.tengo file (depth > 1)
+
+## 1.17.1
+
+### Patch Changes
+
+- a974567: Strict tengo compiler options control
+
 ## 1.17.0
 
 ### Minor Changes